refactor user repo name
