+++ app/src/main/java/com/example/newsapp/data/repos/UserRepo.kt
@@ -1,0 +1,12 @@
+package com.example.newsapp.data.repos
+
+import com.example.newsapp.data.entities.UserEntity
+
+interface UserRepo {
+
+    suspend fun registerUser(userEntity: UserEntity): Boolean
+
+    suspend fun login(email: String, password: String): Boolean
+
+    suspend fun changePassword(email: String, password: String): Boolean
+}
